+++ xpybuild/internal/scheduler.py
@@ -29,7 +29,7 @@ from functools import reduce
 from xpybuild.buildcommon import XPYBUILD_VERSION
 from xpybuild.basetarget import BaseTarget
 from xpybuild.buildcontext import BuildContext
-from xpybuild.buildexceptions import BuildException
+from xpybuild.utils.buildexceptions import BuildException
 from xpybuild.internal.targetwrapper import TargetWrapper
 from xpybuild.internal.threadpool import ThreadPool, Utilisation
 from xpybuild.internal.outputbuffering import outputBufferingManager