+++ xpybuild/targets/zip.py
@@ -26,7 +26,7 @@ from xpybuild.pathsets import PathSet, BasePathSet
 from xpybuild.basetarget import BaseTarget
 from xpybuild.utils.fileutils import mkdir, deleteDir, normLongPath
 from xpybuild.utils.flatten import flatten
-from xpybuild.buildexceptions import BuildException
+from xpybuild.utils.buildexceptions import BuildException
 
 class Zip(BaseTarget):
 	""" Target that creates a zip archive from a set of input files.